set up repo env and added to con cycle algorithm
