+++ .env.example
@@ -1,0 +1,5 @@
+#USE THIS FILE AS A TEMPLATE TO CREATE YOUR OWN .env FILE
+#_______________________________________________________________________________________________________________
+#Condorcet Cycle Variables:
+
+#_______________________________________________________________________________________________________________
