+++ Datasets/README.txt
@@ -1,0 +1,3 @@
+Note: For sake of convenience I am not going to add this path to .gitignore.
+However, if we do end up publishing something, it would be best practice 
+to get permission to include the data we use in our study.
